requirements.txt
fastapi==0.115.6
requests==2.31.0
beautifulsoup4==4.12.3
trafilatura>=1.12.0
lxml>=5.2.2
python-dotenv==1.0.1
pydantic==2.10.4
httpx>=0.27.0
